@jan--f 2023-03-07
Do you have some performance numbers on the speedup you're seeing? I did a `make clean; time make generate` (run multiple times in a row) on master and on your PR with counter intuitive results:
`master`:
```
make generate  32.63s user 4.52s system 65% cpu 56.911 total
make generate  33.86s user 4.96s system 63% cpu 1:01.18 total
```
This PR:
```
make generate  43.42s user 6.48s system 114% cpu 43.555 total
make generate  43.82s user 5.42s system 125% cpu 39.260 total
```