@sthaha 2023-03-08
> Do you have some performance numbers on the speedup you're seeing? I did a `make clean; time make generate` (run multiple times in a row) on master and on your PR with counter intuitive results: `master`:
> 

I don't think `make generate` is the correct target to run since it does more than just generate yaml from jsonnet. Please try
` time -p make build-jsonnet`  which produces the following  on my machine. Are you running this in a container with only one cpu allocated by any chance? 

```
time -p make build-jsonnet

# master 
real 28.69
user 27.60
sys 4.08

# this pr 
real 10.62
user 34.76
sys 5.13

```