@jan--f 2023-03-08
I still get very similar results with `make clean; export TIMEFMT='%E'; time make build-jsonnet`: 
`main`:
```
47.94s 47.37s
```
PR: 
```
45.35s 39.62s
```