@sthaha 2023-03-08
> make clean; export TIMEFMT='%E'; time make build-jsonnet

I think this could be the issue, that `make clean` deletes everything, so the time is actually measuring all the download of tools , jsonnet bundles etc .. E.g.

```
❯ make clean; export TIMEFMT='%E'; time make build-jsonnet
rm -rf jsonnet/vendor operator .hack-operator-image tmp/
mkdir -p /home/sthaha/dev/rh/monitoring/downstream/cmo/wt/improve-jsonnet-gen/tmp/bin
Installing tools from hack/tools.go
go build '-mod=mod' -o cmo/wt/improve-jsonnet-gen/tmp/bin github.com/brancz/gojsontoyaml
go build '-mod=mod' -o cmo/wt/improve-jsonnet-gen/tmp/bin github.com/campoy/embedmd
go build '-mod=mod' -o cmo/wt/improve-jsonnet-gen/tmp/bin github.com/google/go-jsonnet/cmd/jsonnet
go build '-mod=mod' -o cmo/wt/improve-jsonnet-gen/tmp/bin github.com/google/go-jsonnet/cmd/jsonnetfmt
go build '-mod=mod' -o cmo/wt/improve-jsonnet-gen/tmp/bin github.com/jsonnet-bundler/jsonnet-bundler/cmd/jb
go build '-mod=mod' -o cmo/wt/improve-jsonnet-gen/tmp/bin github.com/prometheus/prometheus/cmd/promtool
cd jsonnet && cmo/wt/improve-jsonnet-gen/tmp/bin/jb install
GET https://github.com/prometheus-operator/kube-prometheus/archive/74cb76d338f021bdca9af635c2eef6b196804b57.tar.gz 200
GET https://github.com/prometheus-operator/prometheus-operator/archive/f337e4ecf88d2c228c2325249bf338c6aa999488.tar.gz 200
GET https://github.com/openshift/openshift-state-metrics/archive/7beb8807e2c0092b5e580a29903ff060140d8ff0.tar.gz 200
GET https://github.com/openshift/telemeter/archive/ee1ba4699b82ecb2033f886af12b9e2e451d2296.tar.gz 200
```

Instead could you please try just running `rm -rf tmp/json-manifests; time make build-jsonnet`  multiple times without `make clean`.
The target deletes all assets and manifests, so those are generated every time.