@jan--f 2023-03-08
Yeah that looks more promising:

`master`:
```
real	0m19.456s
user	0m19.835s
sys	0m1.445s
52.00s

real	0m20.376s
user	0m20.634s
sys	0m1.533s
49.92s
```
PR:
```
real	0m8.408s
user	0m31.937s
sys	0m2.293s
38.19s

real	0m8.448s
user	0m31.868s
sys	0m2.275s
39.01s
```