@jan--f 2023-04-03
yes, sorry for the lag!
just did a run again and get consistently numbers like `real    0m19.300s` prior to this change, `real    0m7.420s` with this change.
/lgtm
/label docs-approved
/label px-approved
/label qe-approved
Adding labels for an internal build optimization.